First Night Done
Added "Warmth meters" for the player and cat, which deplete over time.

Modeled a flaming barrel and steam grate, which provide heat when the player is near.

The player can use the "E" key to transfer their body heat to the cat to keep it alive, at the cost of depleting their warmth faster.

Modeled and incorporated the goal/Vet building. When the player reaches the entrance, they win, and the cat is saved.